refactor(desktop): github pr creation as inline inset panel
Replace the CreatePrDialog modal with an inline CreatePrPanel rendered inside
PrDetailPanel, matching the gitlab MR inset-panel design (title/description/base,
draft toggle, draft-with-an-agent).
